# Speedup Llama2 cpu throughput in bench by 1.69x with iobinding