@kunal-vaishnavi 2024-03-11
Should we manually set `use_io_binding = True` anyways for readability purposes and add a comment explaining why it is always true? It might be confusing why `use_io_binding` is not enabled.