@BowenBao 2024-03-11
The other perspective is that `use_io_binding` can be viewed as impl detail of optimum, and we trust optimum to pick the best choice. This way there are less mental burden in writing the benchmark script. But I'm also okay with your proposal. 

I wonder if using iobinding / dlpack is always better, and if we should do something to improve the default `sess.run` with it.